+++ lib/src/index.ts
@@ -5,6 +5,7 @@ export * from './json';
 // Internals (for backend implementations)
 export * from './metadata';
 export * from './types';
+export * from './frontend_options';
 export { default as DecoratorFactory } from './decorator_factory';
 export { default as ObjectPropertySerializer } from './object_property_serializer';
 export { default as PropertySerializer } from './property_serializer';